+++ client/components/dashboard/index.js
@@ -4,6 +4,8 @@ import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as actionCreators from '../../actions/orderActions';
 
+import { cashTransactionTypeId, creditCardTransactionTypeId } from '../../utilities/ordersUtility';
+
 import MonthlySummary from './monthlySummary';
 import SalesWidget from './SalesWidget.jsx';
 import OrdersWidget from './OrdersWidget.jsx';
@@ -15,21 +17,30 @@ import styles from './dashboard.css';
 class Dashboard extends React.Component {
 
   render() {
+    const {
+      yesterdaysSales,
+      todaysSales,
+      yesterdaysCount,
+      todaysCount,
+      todaysCashTransactionSales,
+      todaysCreditCardTransactionSales
+    } = this.props;
+
     return (
       <div className="row">
         <div className="col-lg-12">
           <div className={styles['widget-container']}>
             <SalesWidget
-              currentSales={1304}
-              yesterdaysSales={1500}
+              currentSales={todaysSales}
+              yesterdaysSales={yesterdaysSales}
             />
             <OrdersWidget
-              currentOrders={152}
-              yesterdaysOrders={80}
+              yesterdaysOrders={yesterdaysCount}
+              currentOrders={todaysCount}
             />
             <TransactionsWidget
-              creditCardTransactions={480}
-              cashTransactions={824}
+              creditCardTransactions={todaysCreditCardTransactionSales}
+              cashTransactions={todaysCashTransactionSales}
             />
             <InventoryWidget />
           </div>
@@ -43,12 +54,52 @@ class Dashboard extends React.Component {
 }
 
 Dashboard.propTypes = {
-  orderActions: PropTypes.object.isRequired
+  orderActions: PropTypes.object.isRequired,
+  yesterdaysSales: PropTypes.number.isRequired,
+  todaysSales: PropTypes.number.isRequired,
+  yesterdaysCount: PropTypes.number.isRequired,
+  todaysCount: PropTypes.number.isRequired,
+  todaysCashTransactionSales: PropTypes.number.isRequired,
+  todaysCreditCardTransactionSales: PropTypes.number.isRequired
 };
 
-export function mapStateToProps() {
-  return {};
+export function mapStateToProps(state) {
+
+  const yesterdaysSales = state.orders.yesterdaysOrders
+    .reduce((previous, next) => {
+      return previous + next.total;
+    }, 0);
+
+  const todaysSales = state.orders.todaysOrders
+    .reduce((previous, next) => {
+      return previous + next.total;
+    }, 0);
+
+  const yesterdaysCount = state.orders.yesterdaysOrders.length;
+  const todaysCount = state.orders.todaysOrders.length;
+
+  const todaysCashTransactionSales = state.orders.todaysOrders
+    .filter(order => order.transactionTypeID === cashTransactionTypeId)
+    .reduce((previous, next) => {
+      return previous + next.total;
+    }, 0);
+
+  const todaysCreditCardTransactionSales = state.orders.todaysOrders
+    .filter(order => order.transactionTypeID === creditCardTransactionTypeId)
+    .reduce((previous, next) => {
+      return previous + next.total;
+    }, 0);
+
+  return {
+    yesterdaysSales,
+    todaysSales,
+    yesterdaysCount,
+    todaysCount,
+    todaysCashTransactionSales,
+    todaysCreditCardTransactionSales
+  };
 }
+
 function mapDispatchToProps(dispatch) {
   return {
     orderActions: bindActionCreators(actionCreators, dispatch)